r tar i bruk onboaring-komponent i ytelser-onboarding
